Add files via upload
This is my first dashboard which I created using Power BI. I utilized Power Pivot, DAX and Powe Query to create data models, perform calculations and establish relationships. This dashboard was my attempt to enter the #mavenpizzachallenge by Maven Analytics which is the source of the data. I have performed analysis of the transactional data of Plato's Pizza by creating charts and tables to visualize and understand the pizza sales related information and to create new business ideas from the insights. There is a lot of room for improvement which I would try to cover up in future projects.